refactor: Update endpoints to use POST requests
Update several GET endpoints to POST for enabling/disabling,
setting memory, latency, window size, percentile, wait time, and
resetting the breaker.  This change aligns better with RESTful
practices for modifying server state. The example server is also
updated to reflect the change from GET to POST. Also, the breaker
is instantiated with a pointer of config.